lyricflow/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-multipart==0.0.12
jinja2==3.1.4
aiofiles==24.1.0
python-dotenv==1.0.1
anthropic==0.39.0
openai-whisper==20240930
rapidfuzz==3.14.3
pytest==8.3.4
httpx==0.27.2
